z-depth working. still rendering upside down

+++ video/pingo/render/depth.h
@@ -7,7 +7,8 @@
 
 #ifdef ZBUFFER32
 typedef struct tag_PingoDepth {
-    uint32_t d;
+    // uint32_t d;
+    float d;
 } PingoDepth;
 #endif
